requirements.txt
Flask
Pillow>=8.3.1
Flask-CORS==3.0.10
rembg
werkzeug
